Change the data type of atk and blood(include max_blood) from int to float to insure correct blood dropping.

+++ Classes/Properties.h
@@ -6,10 +6,10 @@ USING_NS_CC;
 
 class PropertiesBundle{
 public:
-	int speed;
+	float speed;
 	int hitBoxRadius;
-	int atk; // drop atk per second
-	int max_blood;
+	float atk; // drop atk per second
+	float max_blood;
 };
 
 class normalZombieProp : public PropertiesBundle {
